Implement queue subscriber against Reactive Streams TCK.
UnicastSubscriber is based on the Reactive Streams example, and
catches the various rules we missed the first time around.  A
big difference is that we expose a method for requesting,
instead of requesting every time we receive (thus blocking
the producer when inserting into a bounded queue with a slow
consumer).

QueueSubscriber extends that to generically adapt a Publisher
to a Process.

The client extends the new QueueSubscriber to do most of the
dirty work.